refactor: fold single-consumer TUI/container modules into their importers (#2858)
