+++ src/klangk/klangk/cli/tui/screens/workspace_detail.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import re
 import asyncio
 import logging
 import subprocess
@@ -26,7 +27,6 @@ from textual.widgets import (
 )
 
 from ...client import AuthError, WorkspaceNotFoundError
-from ..marking import effective_marking, marking_style
 from ._base import (
     CheatsheetScreen,
     ConfirmScreen,
@@ -1347,3 +1347,52 @@ class WorkspaceDetailScreen(StatusScreen):
         else:
             # payload is the error text; show it inline on the detail screen.
             self._msg(payload, error=True)
+
+
+# ---------------------------------------------------------------------------
+# Classification-marking rendering (#2768; moved verbatim from the former
+# tui/marking submodule — sole consumer was this screen).
+# ---------------------------------------------------------------------------
+
+# (pattern, background) pairs checked in order against the marking; the
+# first hit wins (TOP SECRET must precede SECRET). Word-boundary matching
+# so a free-text label that merely *contains* a marking word (e.g.
+# "NOT SECRETIVE") is not colored as that marking.
+_MARKING_COLORS: tuple[tuple[re.Pattern[str], str], ...] = tuple(
+    (re.compile(rf"\b{word}\b", re.IGNORECASE), color)
+    for word, color in (
+        ("TOP SECRET", "#E0A800"),
+        ("SECRET", "#C01818"),
+        ("CONFIDENTIAL", "#005EB8"),
+        ("CUI", "#0076CE"),
+        ("UNCLASSIFIED", "#007A33"),
+    )
+)
+_FALLBACK_COLOR = "#8A6D00"  # neutral amber for free-text labels
+
+
+def marking_background(marking: str) -> str:
+    """The banner background color (#rrggbb) for a marking label."""
+    for pattern, color in _MARKING_COLORS:
+        if pattern.search(marking):
+            return color
+    return _FALLBACK_COLOR
+
+
+def marking_style(marking: str) -> str:
+    """A Rich style string for rendering a marking line (white on color)."""
+    return f"bold white on {marking_background(marking)}"
+
+
+def effective_marking(
+    workspace_banner: str | None, deploy_default: str | None
+) -> str:
+    """Resolve the marking actually shown: workspace override, else deploy.
+
+    Whitespace-only values count as unset; the result is stripped, and an
+    empty resolution means "render nothing" (no reserved screen space).
+    """
+    own = (workspace_banner or "").strip()
+    if own:
+        return own
+    return (deploy_default or "").strip()
