+++ src/klangk/klangk/cli/tui/screens/workspace_form.py
@@ -26,7 +26,6 @@ from textual.widgets import (
 from textual.widgets.option_list import Option
 
 from ...client import AuthError, Workspace
-from ...env import validate_env_entry
 from ...mount import (
     validate_allowed_domain_spec,
     validate_mount_spec,
@@ -1661,3 +1660,25 @@ class EditWorkspaceScreen(WorkspaceFormMixin, TabSkipMixin, StatusScreen):
             "pids_limit",
         ):
             self._save()
+
+
+# ---------------------------------------------------------------------------
+# Client-side env-entry validation (moved verbatim from the former cli/env
+# submodule — sole consumer was this screen).
+# ---------------------------------------------------------------------------
+
+
+def validate_env_entry(spec: str) -> str | None:
+    """Validate a ``KEY=VALUE`` environment variable entry.
+
+    Returns None if valid, or an error message string if invalid.
+    Mirrors the Flutter ``CreateWorkspaceDialog`` rule: the entry must
+    contain ``=`` and have a non-empty key (the part before the first
+    ``=``). The value may be empty.
+    """
+    if "=" not in spec:
+        return f"Invalid env {spec!r}: expected KEY=VALUE"
+    key, _, _ = spec.partition("=")
+    if not key:
+        return f"Invalid env {spec!r}: key cannot be empty"
+    return None
